renamed springSecurity into security, moved panelAuth into security folder
