+++ src/main/talium/system/security/HeaderAuthProcessingFilter.java
@@ -1,11 +1,11 @@
-package talium.system.springSecurity;
+package talium.system.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import talium.system.panelAuth.AuthService;
-import talium.system.panelAuth.AuthenticationRejected;
+import talium.system.security.auth.AuthService;
+import talium.system.security.auth.AuthenticationRejected;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
