+++ src/main/talium/system/security/auth/persistence/Session.java
@@ -1,4 +1,4 @@
-package talium.system.panelAuth.persistence;
+package talium.system.security.auth.persistence;
 
 import jakarta.persistence.*;
 import org.jetbrains.annotations.NotNull;
