+++ src/main/talium/system/security/auth/persistence/SessionRepo.java
@@ -1,4 +1,4 @@
-package talium.system.panelAuth.persistence;
+package talium.system.security.auth.persistence;
 
 import org.springframework.data.repository.CrudRepository;
